[WIP] big refactor
