+++ README.md
@@ -1,4 +1,3 @@
 # dex_stats_pymongo
-mongodb swap status jsons parser
 
-aim is to provide all required data for [CMC "Ideal endpoint"](https://docs.google.com/document/d/1S4urpzUnO2t7DmS_1dc4EL4tgnnbTObPYXvDeBnukCg/edit#)
+The aim is to provide all required data for [CMC "Ideal endpoint"](https://docs.google.com/document/d/1S4urpzUnO2t7DmS_1dc4EL4tgnnbTObPYXvDeBnukCg/edit#)
