+++ dex_stats/ParserError.py
@@ -1,0 +1,9 @@
+class Parser_Error(Exception):
+    pass
+
+
+
+class ArgumentInputParserError(Parser_Error):
+    def __init__(self, expression, message):
+        self.expression = expression
+        self.message = message
