+++ dex_stats/utils/adex_calls.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python3
 import requests
 
 
@@ -6,10 +5,13 @@ def batch_request(node_ip, user_pass, requests_list):
     r = requests.post(node_ip, json=requests_list)
     return r
 
+
 def get_orderbook(node_ip, user_pass, base, rel):
-    params = {'userpass': user_pass,
+    params = {
+              'userpass': user_pass,
               'method': 'orderbook',
-              'base': base, 'rel': rel,}
+              'base': base,
+              'rel': rel
+             }
     r = requests.post(node_ip, json=params)
     return r
-
