+++ dex_stats/utils/enable_coins.py
@@ -3,4 +3,3 @@ from adex_calls import batch_request
 
 batch_request("http://127.0.0.1:7783", "testuser", enable_calls)
 batch_request("http://127.0.0.1:7783", "testuser", electrum_calls)
-
